+++ inst/test_backends/test_MsBackend/test_spectra_subsetting.R
@@ -59,6 +59,19 @@ test_that("[", {
     expect_equal(res, be[which(l)])
 })
 
+test_that("cbind2 works", {
+    seql <- length(be)
+    df <- data.frame(cola = seq_len(seql), colb = "b", colz = "z")
+    res <- cbind2(be, df)
+    expect_true(validObject(res))
+    expect_equal(ncol(spectraData(res)), length(spectraVariables(be)) + 3)
+    expect_equal(res$cola, seq_len(seql))
+    expect_equal(res$colb, rep("b", seql))
+    expect_equal(res$colz, rep("z", seql))
+    df2  <- data.frame(cola = 3:6, colb = "b", colz = "z")
+    expect_error(cbind2(be, df2), "does not match")
+})
+
 #' extractByIndex. Uses [ if not implemented
 test_that("extractByIndex", {
     i <- sample(seq_along(be), floor(length(be) / 2))